requirements.txt
torch>=2.1
transformers>=4.40
sentencepiece>=0.1.99
accelerate>=0.33
datasets>=2.19
pymupdf>=1.24.8
pdfminer.six>=20240706
nltk>=3.9
rouge-score>=0.1.2
sacrebleu>=2.4.2
streamlit>=1.36
tqdm>=4.66
numpy>=1.26
